+++ DQMUtils/DQMHistogramSignalScanFit.cc
@@ -35,10 +35,9 @@ DQMHistogramSignalScanFit::~DQMHistogramSignalScanFit ()
 
 
 //========================================================================================================================
-void DQMHistogramSignalScanFit::book(TFile *theOutputFile, const DetectorContainer &theDetectorStructure, std::map<std::string, uint32_t> pSettingsMap)
+void DQMHistogramSignalScanFit::book(TFile *theOutputFile, const DetectorContainer &theDetectorStructure, std::map<std::string, double> pSettingsMap)
 {
     ContainerFactory::copyStructure(theDetectorStructure, fDetectorData);
-
 }
 
 //========================================================================================================================
