+++ DQMUtils/DQMHistogramTPCalibration.cc
@@ -35,7 +35,7 @@ DQMHistogramTPCalibration::~DQMHistogramTPCalibration ()
 
 
 //========================================================================================================================
-void DQMHistogramTPCalibration::book(TFile *theOutputFile, const DetectorContainer &theDetectorStructure, std::map<std::string, uint32_t> pSettingsMap)
+void DQMHistogramTPCalibration::book(TFile *theOutputFile, const DetectorContainer &theDetectorStructure, std::map<std::string, double> pSettingsMap)
 {
     ContainerFactory::copyStructure(theDetectorStructure, fDetectorData);
